requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
crispy-bootstrap5==2026.3
cryptography==46.0.7
cssselect2==0.9.0
Django==6.0.4
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-filter==25.2
djangorestframework==3.17.1
django-redis==6.0.0
redis==7.4.0
psycopg==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.2.2
fonttools==4.62.1
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.4
oscrypto==1.3.0
pillow==12.2.0
pycairo==1.29.0
pycparser==3.0
pydyf==0.12.1
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
pypdf==6.10.2
pyphen==0.17.2
python-bidi==0.6.7
python-dotenv==1.2.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.33.1
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2026.1
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.4.1
openpyxl==3.1.3
